requirements.txt
wget
tgcrypto
ffmpeg-python
py-tgcalls==0.9.7
telethon==1.28.2
pyrogram
youtube-search
youtube-search-python
yt-dlp
python-dotenv
dnspython
gitpython
asgiref
asSQL
numpy
unidecode
getids
aiofiles
dragonxxdlib
getids
SpeechRecognition
aiohttp
requests
pillow==9.5.0
heroku3
psutil
future
wget
speedtest-cli
motor
pyromod~=1.5
googletrans
pytz
